Fix broadcast receiver register on android 14.0/34+
This adds a new helper method for registering broadcast receivers, which takes a bool for exporting the service or not, and on API 34 and higher, it calls the newer overload which takes the `ReceiverFlags` and passes the corresponding `NotExported` or `Exported` value to the bool parameter of the util method.

This allows us to call the right overload if we are on API 34 (Android 14.0) or newer where the declaration is required.

Also some older GetLocale()/SetLocale() methods are removed which were unused.